feat(restapi): pushing for debug sesh
